qa-watchdog/requirements.txt
# QA Watchdog Bot Dependencies

# Telegram
python-telegram-bot>=21.0

# Gemini AI
google-genai>=1.0.0

# Scheduler
APScheduler>=3.10.0

# Web server for health checks
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
